requirements.txt
matplotlib
numpy
omegaconf
opencv
pandas
Pillow
anaconda::scikit-learn
pytorch::pytorch
pytorch::torchvision
pytorch::pytorch-cuda
torchvision
tqdm
wandb